Remove the current-checkout prefix from the load path before testing it.
Without this change, a checkout of bundler to "bundler-0.9" (wherever in
the absolute path) will avoid the check for older bundler versions (as it
hits the _first_ bundler-0\.\d in the path.

With this change, the test (and the failsafe as well) are made more solid,
if not yet 100% perfect.